requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
python-dotenv==1.0.1
azure-storage-blob==12.25.0
<<<<<<< HEAD
mysql-connector-python==9.1.0
=======
mysql-connector-python==9.1.0
>>>>>>> origin/main
